+++ AIs/GeneralAI/ExampleAI/ExampleAI.h
@@ -9,8 +9,8 @@ class ExampleAI : public GeneralAI
         ExampleAI(int inputSize, int outputSize, int maxInput, int maxOutput);
 
     private:
-        std::vector<int> coreOutput(const std::vector<int>& input);
-        void coreLearn(const std::vector<int>& input, const std::vector<int>& output, float outcome);
+        std::vector<float> coreOutput(const std::vector<float>& input);
+        void coreLearn(const std::vector<float>& input, const std::vector<float>& output, float outcome);
 
         template <class Archive>
         void serialize( Archive & ar )
